requirements1.txt
# local package
-e .

# external requirements
black
click==8.1.3
fastapi==0.89.1
flake8
hydra-core==1.3.1
isort
loguru==0.6.0
Markdown==3.4.1
matplotlib==3.6.2
mypy
numpy==1.23.4
omegaconf==2.3.0
opencv_python==4.7.0.68
protobuf==4.21.12
pydantic==1.10.4
pytorch_lightning
PyYAML
scikit_learn==1.2.0
setuptools==65.6.3
timm==0.6.12
torch==1.13.0
torchdrift
torchvision
uvicorn
wandb==0.13.8
